BaseListeningAgent as parent for Agent
