+++ examples/examples_with_threads/test_listenning_agent.metta
@@ -5,6 +5,6 @@
 ! (&a1 .input "who is the 6 president of France")
 ! ((py-atom time.sleep) 2)
 ! (&a1 .input "who is John Lennon?")
-;! (&a1 .cancel_processing)
+! (&a1 .cancel_processing)
 ! ((py-atom time.sleep) 3)
 ! (&a1 .stop)
